Use long int for file sizes and explicitly convert to int where needed

+++ DFileUtils.h
@@ -19,11 +19,11 @@ char *get_leaf_name(DFile *dfile);
 SFError save_compressed(DFile *dfile, char *fname);
 SFError load_compressed(DFile *dfile, char const *fname);
 
-int get_compressed_size(DFile *dfile);
+long int get_compressed_size(DFile *dfile);
 
 SFError write_compressed(DFile *dfile, struct Writer *writer);
 SFError read_compressed(DFile *dfile, struct Reader *reader);
 
-int worst_compressed_size(DFile *dfile);
+long int worst_compressed_size(DFile *dfile);
 
 #endif
